fix: use current user data for email verification
- fixes an update bug with where submitRegistrationMetadata was updating
the database but not the in-memory User object that still does not have
an actual email before calling sendEmailVerification. This led to the
first submission not sending an email out, but subsequent form
submissions properly sending the email out.

- push sendEmailVerification into submitRegistrationMetadata and ensure
that it is operating on a User object

- replace existingUser in consent form with a getter that checks
the status of the dateConsented field

Co-authored-by: Christine Nguyễn <chrstngyn@gmail.com>
Co-authored-by: Calvin Pritchard <pritchard.calvin@gmail.com>