refactor: accomplishment cards emit discard event
parent components are responsible for handling the discard event and
invoking the proper API method to discard the accomplishment card

part of this is to visually refactor the efforts wasted event (for #609)
to use the same AccomplishmentCard UI to render the accomplishment cards
that a player must select to discard. still need to deal with how to
properly display multiple purchased accomplishments to discard though.

Co-authored-by: Calvin Pritchard <pritchard.calvin@gmail.com>
Co-authored-by: Christine Nguyễn <chrstngyn@gmail.com>