fix(solo): prevent investing before round init and persist before gameover
If initialization was slow enough, the investment command could be
queued before the round was fully set up, which caused compounding
inconsistencies. We can use the canInvest flag to prevent this

After moving system health deduction to the end of the round, the game
would end before persisting the round if caused by the natural system
health deduction. Now a persist cmd is queued up before the end cmd